requirements.txt
playwright
pytest
pytest-playwright
allure-pytest
pandas
openpyxl
sweetviz
ydata-profiling
snowflake-connector-python
